+++ hub/core/linked_sample.py
@@ -1,5 +1,13 @@
 from typing import Optional
-from hub.util.link import convert_creds_key
+from hub.constants import ALL_CLOUD_PREFIXES
+
+
+def convert_creds_key(creds_key: Optional[str], path: str):
+    if creds_key is None and path.startswith(ALL_CLOUD_PREFIXES):
+        creds_key = "ENV"
+    elif creds_key == "ENV" and not path.startswith(ALL_CLOUD_PREFIXES):
+        creds_key = None
+    return creds_key
 
 
 class LinkedSample:
